Move User class
